refactor luggage all Luggage et test

+++ flycheckin/boarding/src/main/java/ca/ulaval/glo4002/flycheckin/boarding/domain/luggage/CarryOnLuggage.java
@@ -9,7 +9,7 @@ public class CarryOnLuggage extends Luggage {
   private static final String TYPE = "carry-on";
   private static final int BASE_PRICE = 30;
 
-  public CarryOnLuggage(int linearDimension, int weight) {
+  public CarryOnLuggage(double linearDimension, double weight) {
     super(linearDimension, weight);
   }
